requirements.txt
asgiref==3.5.2
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.12
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.14.0
django-debug-toolbar==1.10.1
django-utils-six==2.0
gunicorn==20.1.0
idna==2.8
oauthlib==3.0.1
pep8==1.7.1
Pillow==9.2.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2022.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.16.0
sqlparse==0.4.2
stripe==2.27.0
tzdata==2022.1
urllib3==1.24.2
